# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZIP 042 API",
    "version": "rules-2026-08-20.1 / extraction-2026-08-20",
    "description": "사회초년생을 위한 원룸 계약 안전 검사 서비스 백엔드.\n\n## 인증\n`Authorization: Bearer <Supabase access token>`\n목 모드(`ZIP042_MODE=mock`)에서는 아무 문자열이나 보내면 되고, 토큰이 곧 사용자 식별자입니다.\n\n## 시작하기\n1. `GET /v1/meta` — 기능 가용성 확인 (시세 조회 가능 여부 등)\n2. `GET /v1/addresses/search` — 주소를 골라 좌표·법정동코드를 얻습니다 (**필수**)\n3. `POST /v1/cases` — 검사 건 생성\n4. `POST /v1/cases/{caseId}/documents/upload-url` → PUT → `POST .../documents`\n5. `POST /v1/cases/{caseId}/analyze/jobs` — 비동기 분석 (권장) 또는 `POST .../analyze` (동기)\n\n## 판정의 두 축\n`verdict.score` 는 **실제 위험**만 반영합니다. 확인하지 못한 항목은 `informationGaps` 로\n따로 나가고, 그중 판정을 완료할 수 없게 만드는 것은 `blockingGaps` 입니다.\n`blockingGaps` 가 비어 있지 않으면 `contractable` 은 false 이며, 이때 UI 는\n\"위험합니다\"가 아니라 **\"아직 판단할 수 없습니다\"** 로 표현해야 합니다.\n\n## 주의\n모든 판정 응답에는 `caveats` / `disclaimer` 가 포함됩니다. **법률 자문이 아니라는 고지를**\n**UI 에 반드시 노출해야 합니다.**"
  },
  "servers": [
    {
      "url": "http://localhost:8787"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "UNPROCESSABLE",
                  "VALIDATION_FAILED",
                  "RATE_LIMITED",
                  "UPSTREAM_FAILED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "사용자에게 그대로 보여줄 수 있는 한국어 메시지"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "UploadUrlRequest": {
        "type": "object",
        "properties": {
          "docType": {
            "type": "string",
            "enum": [
              "registry",
              "brokerage_statement",
              "lease_draft",
              "building_ledger",
              "other"
            ]
          },
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "application/pdf",
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        },
        "required": [
          "docType",
          "fileName",
          "mimeType"
        ],
        "additionalProperties": false
      },
...
      "AnalyzeRequest": {
        "type": "object",
        "properties": {
          "refreshMarketPrice": {
            "type": "boolean",
            "default": false
          },
          "reparseDocuments": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 목록",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cases": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Case"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 생성",
        "description": "금액은 `amountUnit` 으로 단위를 명시합니다(`krw` 또는 `man`). 좌표·법정동코드를 비워두면 서버가 주소로 보완을 시도합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "case": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{caseId}/documents/upload-url": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "서명 업로드 URL 발급 (1/3단계)",
        "description": "20MB 파일을 API 서버로 통과시키지 않기 위해 클라이언트가 Storage 로 직접 PUT 합니다. 한도: 30회/시간.",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "발급됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "upload": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string",
                          "description": "이 URL 로 파일 본문을 PUT 하세요"
                        },
                        "storagePath": {
                          "type": "string",
                          "description": "3단계에서 그대로 넘깁니다"
                        },
                        "method": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/cases/{caseId}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "분석 실행 (동기)",
        "description": "문서 판독이 포함되면 **20~60초**가 걸릴 수 있습니다. 타임아웃을 90초 이상 두세요. 한도: 10회/시간. 서버리스 환경에서는 이 경로를 쓰세요.",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "분석 결과",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "analysis": {
                      "type": "object",
                      "properties": {
                        "version": {
                          "type": "integer"
                        },
                        "verdict": {
                          "$ref": "#/components/schemas/Verdict"
                        },
                        "findings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Finding"
                          }
                        },
                        "specialTerms": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SpecialTerm"
                          }
                        },
                        "schedule": {
                          "type": "object",
                          "properties": {
                            "evaluable": {
                              "type": "boolean"
                            },
                            "opposingPowerEffectiveAt": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true,
                              "description": "대항력 발생 시점 (전입신고 다음 날 0시 KST)"
                            },
                            "priorityRightEffectiveAt": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true
                            },
                            "unprotectedWindow": {
                              "type": "object",
                              "nullable": true,
                              "description": "잔금 지급 후 법적 보호를 받지 못하는 구간",
                              "properties": {
                                "fromDate": {
                                  "type": "string",
                                  "format": "date"
                                },
                                "toDate": {
                                  "type": "string",
                                  "format": "date"
                                },
                                "days": {
                                  "type": "integer"
                                }
                              }
                            },
                            "events": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/ScheduleEvent"
                              }
                            }
                          }
                        },
                        "valuation": {
                          "type": "object",
                          "description": "시세 · 부담률 · 경매 회수 시뮬레이션"
                        },
                        "region": {
                          "type": "object",
                          "description": "지역 위험"
                        },
                        "crossCheck": {
                          "type": "object",
                          "description": "서류 교차검증"
                        },
                        "caveats": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "판정의 한계. UI 에 반드시 노출하세요."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{caseId}/analyze/jobs": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "분석 실행 (비동기, 권장)",
        "description": "202 를 즉시 반환하고 백그라운드에서 분석합니다. 모바일 네트워크가 끊겨도 결과를 잃지 않습니다.\n\n1. `POST` → `{ job: { id, status } }`\n2. `GET .../jobs/{jobId}` 로 2.5초 간격 폴링\n3. `status === \"succeeded\"` → `GET .../analysis`\n\n진행 중 작업이 있으면 새로 만들지 않고 그 작업을 돌려줍니다(멱등, 200 + `reused: true`).",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "진행 중 작업 재사용",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "202": {
            "description": "작업 등록됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": {
                      "$ref": "#/components/schemas/AnalysisJob"
                    },
                    "reused": {
                      "type": "boolean"
                    },
                    "poll": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "작업 목록",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/checklist/progress": {
      "get": {
        "tags": [
          "checklist"
        ],
        "summary": "내 체크 상태",
        "description": "한 번도 저장하지 않았으면 빈 상태를 돌려줍니다 (404 아님).",
        "responses": {
          "200": {
            "description": "체크 상태",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklist": {
                      "$ref": "#/components/schemas/ChecklistState"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "checklist"
        ],
        "summary": "체크 상태 저장 (전체 교체)",
        "description": "부분 갱신이 아니라 전체 교체입니다. 화면이 가진 체크 목록 전체를 보내세요. 알 수 없는 항목 id 는 서버가 조용히 걸러냅니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "checkedItemIds"
                ],
                "properties": {
                  "checkedItemIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "저장된 상태",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklist": {
                      "$ref": "#/components/schemas/ChecklistState"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "checklist"
        ],
        "summary": "체크 전체 해제",
        "responses": {
          "200": {
            "description": "초기화된 상태",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklist": {
                      "$ref": "#/components/schemas/ChecklistState"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}